backend/requirements.txt
asgiref==3.11.0
attrs==25.4.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
click-plugins==1.1.1.2
cligj==0.7.2
colorama==0.4.6
contourpy==1.3.3
coreapi==2.3.3
coreschema==0.0.4
cycler==0.12.1
Django==6.0
django-filter==25.2
djangorestframework==3.16.1
drf-yasg==1.21.11
fiona==1.10.1
fonttools==4.61.1
geopandas==1.1.2
idna==3.11
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.8
numpy==2.4.0
packaging==25.0
pandas==2.3.3
pillow==12.0.0
pyogrio==0.12.1
pyparsing==3.2.5
pyproj==3.7.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
shapely==2.1.2
six==1.17.0
sqlparse==0.5.4
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.6.3
